+++ internal/vo/bot.go
@@ -2,7 +2,7 @@ package vo
 
 // 机器人配置数据结构
 type BotConfigVo struct {
-	Id               int64  `json:"id"`
+	Id               uint   `json:"id"`
 	Region           string `json:"region"`
 	Name             string `json:"name"`                 // 机器人名称
 	Token            string `json:"token"`                // 机器人token
@@ -14,7 +14,7 @@ type BotConfigVo struct {
 
 // 机器人配置数据结构
 type BotConfigListVo struct {
-	Id              int64  `json:"id"`
+	Id              uint   `json:"id"`
 	Region          string `json:"region"`
 	Name            string `json:"name"`            // 机器人名称
 	GroupID         int64  `json:"groupId"`         // 群组ID
